First functioning version with fake, local, sample data
- Fix job and worker queries
- Change workerId data type from String to Long
- Override the setter method of Job's status property to ensure active is updated anytime the status changes
- Redefine JobStatus.NEW as an inactive status
- Make all Customer and Location properties non-null
- Insert fake Job data into Room upon onNetworkCallError of the getJobs request to temporarily show data in the app until a server can be spun up
- Add a JobResponseFactory for quickly creating fake Jobs
- Update job item view so that driver name and ETA show up in the right-hand column of each card
- Change vector asset fill colors to white
- Fix city-state format resource
- Add quantity string itmes to ETA

+++ app/src/main/java/com/nerdery/rtaza/mvvmdemo/ui/util/TimeUtil.kt
@@ -14,9 +14,9 @@ class TimeUtil {
          * @param promisedEta the promised ETA in milliseconds.
          * @return the difference between the [promisedEta] and the current time in minutes.
          */
-        fun calculateDueInTime(promisedEta: Long): Long {
-            return TimeUnit.MILLISECONDS.toMinutes(promisedEta) -
-                    TimeUnit.MILLISECONDS.toMinutes(getCurrentTimeMillis())
+        fun calculateDueInTime(promisedEta: Long): Int {
+            return (TimeUnit.MILLISECONDS.toMinutes(promisedEta) -
+                    TimeUnit.MILLISECONDS.toMinutes(getCurrentTimeMillis())).toInt()
         }
 
         /**
@@ -26,10 +26,10 @@ class TimeUtil {
          * @return the difference between the current time and the [promisedEta] in minutes.
          * Returns 0 if the difference is negative (not late).
          */
-        fun calculateMinutesLate(promisedEta: Long): Long {
+        fun calculateMinutesLate(promisedEta: Long): Int {
             val diff = TimeUnit.MILLISECONDS.toMinutes(getCurrentTimeMillis()) -
                     TimeUnit.MILLISECONDS.toMinutes(promisedEta)
-            return if (diff > 0) diff else 0
+            return if (diff > 0) diff.toInt() else 0
         }
     }
 }
